requirements.txt
azure-common==1.1.28
azure-core==1.24.1
azure-storage-common==2.1.0
azure-storage-queue==12.3.0
Flask==2.1.2
requests==2.27.1
